requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.2.0
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
